% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mesh_embeddings.R
\name{data_mesh_embeddings}
\alias{data_mesh_embeddings}
\title{Load mesh embeddings and thesaurus from Git}
\usage{
data_mesh_embeddings()
}
\value{
A matrix of mesh embeddings.
}
\description{
Load mesh embeddings and thesaurus from Git
}
