% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mesh_embeddings.R, R/data_mesh_thesaurus.R
\name{data_mesh_thesuarus}
\alias{data_mesh_thesuarus}
\title{Load mesh thesaurus from Git}
\usage{
data_mesh_thesuarus()

data_mesh_thesuarus()
}
\value{
A data frame.

A data frame.
}
\description{
Load mesh thesaurus from Git
}
