% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pharm_action.R
\name{data_pharm_action}
\alias{data_pharm_action}
\title{Load pharmacological actions from Git}
\usage{
data_pharm_action()
}
\value{
A data frame.
}
\description{
Load pharmacological actions from Git
}
