% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pmc_list.R
\name{data_pmc_list}
\alias{data_pmc_list}
\title{Download/load list of PMC full text articles.}
\usage{
data_pmc_list(force_install = F)
}
\value{
A data frame.
}
\description{
Download/load list of PMC full text articles.
}
