% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.clean_nas}
\alias{.clean_nas}
\title{Clean Missing or Invalid Values in Data}
\usage{
.clean_nas(x)
}
\arguments{
\item{x}{A vector that may contain missing or invalid values represented in various formats.}
}
\value{
A vector with standardized missing values represented as `NA`.
}
\description{
This function standardizes the representation of missing or invalid values in data by replacing specific character representations of missing data (' ', 'NA', 'n/a', 'n/a.') with R's standard `NA`.
}
\keyword{internal}
