% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_pubmed.R
\name{.extract_annotations}
\alias{.extract_annotations}
\title{Extract Annotations from PubMed Records}
\usage{
.extract_annotations(g)
}
\arguments{
\item{g}{An XML node set representing a single PubMed record.}
}
\value{
A data frame with annotations extracted from a PubMed record.
}
\description{
Parses XML records from PubMed to extract annotations such as MeSH terms, chemical names, and keywords.
}
\keyword{internal}
