% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_icites.R
\name{.fetch_icites}
\alias{.fetch_icites}
\title{Internal: Fetch Data from iCite Database}
\usage{
.fetch_icites(x)
}
\arguments{
\item{x}{A vector of PubMed IDs for which data is to be fetched from the iCite database.}
}
\value{
A data.frame consisting of the data retrieved from iCite, formatted as CSV.
}
\description{
This internal function is designed to scrape data from the iCite database, a bibliometric tool provided by the NIH. It constructs a URL to query iCite with specified PubMed IDs and retrieves citation metrics and other related data.
}
\keyword{internal}
