% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fetch_records}
\alias{.fetch_records}
\title{Fetch Batch of PubMed Records as XML}
\usage{
.fetch_records(x, sleep)
}
\arguments{
\item{x}{A vector of PubMed record identifiers to be fetched.}
}
\value{
A character string with XML content of PubMed records, or an error object in case of failure.
}
\description{
This function attempts to fetch batches of PubMed records in XML format. It retries multiple times in case of failures.
}
\keyword{internal}
