% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_affiliations.R
\name{.get_affiliations}
\alias{.get_affiliations}
\title{Internal: Extract Author Affiliations from PubMed Records}
\usage{
.get_affiliations(x, sleep)
}
\arguments{
\item{x}{A character vector with search terms or IDs for fetching records from PubMed.}
}
\value{
A data.table consisting of PubMed IDs, author names, and their affiliations.
}
\description{
Function queries PubMed to extract author affiliations from the fetched records. It processes XML records to obtain detailed information about authors, including their names and affiliations.
}
\keyword{internal}
