% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_icites.R
\name{.get_icites}
\alias{.get_icites}
\title{Process and Structure Data from iCite}
\usage{
.get_icites(x)
}
\arguments{
\item{x}{A vector of PubMed IDs for which data has been fetched from the iCite database.}
}
\value{
A data.table enhanced with citation network information and cleaned reference and citation data.
}
\description{
Function processes and structures the data obtained via `.fetch_icites`.
}
\keyword{internal}
