% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_pmc.R
\name{.get_pmc}
\alias{.get_pmc}
\title{Scrape Full Text Entries from PubMed Central (PMC)}
\usage{
.get_pmc(x, sleep)
}
\arguments{
\item{x}{A vector of PMC identifiers for which full-text articles are to be retrieved.}
}
\value{
A data.table with columns for document ID, PMC identifier, section titles, and text content of each section.
}
\description{
This function retrieves full-text articles from PMC using provided PMC identifiers. It downloads and parses XML files to extract article sections and their corresponding text.
}
\keyword{internal}
