% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_pubtations.R
\name{.get_pubtations}
\alias{.get_pubtations}
\title{Extract Records from PubMed's PubTator Tool}
\usage{
.get_pubtations(x, sleep)
}
\arguments{
\item{x}{A vector of PubMed IDs for which annotations are to be retrieved from PubTator.}
}
\value{
A data.table, or NA if no data is available, with columns for PubMed ID, title or abstract location, annotation text, start and end positions of annotations, and annotation types.
}
\description{
This function retrieves annotated bibliographic data from PubMed's PubTator tool. It fetches data using PubMed IDs and processes the JSON response into a structured format.
}
\keyword{internal}
