% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_pubmed.R
\name{.get_records}
\alias{.get_records}
\title{Get PubMed Records}
\usage{
.get_records(x)
}
\arguments{
\item{x}{A character vector with search terms or IDs for fetching records from PubMed.}
}
\value{
A data.table with columns for PubMed IDs, publication year, journal name, article title, abstract, and annotations.
}
\description{
Processes XML records obtained from PubMed. It extracts basic bibliographic information and annotations for each record.
}
\keyword{internal}
