% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pmid_metadata.R
\name{download_pmid_metadata}
\alias{download_pmid_metadata}
\title{Downloads metadata from Pubmed API for provided PMID's}
\usage{
download_pmid_metadata(df, col, api_key)
}
\arguments{
\item{df}{A dataframe containing a column of PMID's}

\item{col}{The name of the column containing PMID's}

\item{api_key}{A valid Pubmed API key}
}
\value{
A data frame containing the original columns as well as
five additional columns:

\if{html}{\out{<div class="sourceCode">}}\preformatted{The `pubmed_dl_success` column is TRUE in the case that
metadata were successfully downloaded from Pubmed; FALSE in the
case that an error occurred during downloading (e.g. due to a
number that is well-formed but does not correspond to a true
PMID); NA in the case that the supplied PMID is not well-formed
(e.g. NA or non-numeric).

The `doi` column returns a DOI that corresponds to the PMID
supplied if one is found, NA otherwise.

The `languages` column contains a JSON-encoded list of
languages for the article in question.

The `pubtypes` column contains a JSON-encoded list of
publication types for the article in question.

The `authors` column contains a JSON-encoded list of authors
for the article in question.
}\if{html}{\out{</div>}}
}
\description{
Downloads metadata from Pubmed API for provided PMID's
}
