% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmed_crosstab_query.R
\name{pmed_crosstab_query}
\alias{pmed_crosstab_query}
\title{Cross-tab search results for multiple calls to pmtk_search_pubmed().}
\usage{
pmed_crosstab_query(x, remove_duplicates = T)
}
\arguments{
\item{x}{A df or list of dfs returned from pmtk_search_pubmed()}

\item{remove_duplicates}{Boolean}
}
\value{
A data frame
}
\description{
Cross-tab search results for multiple calls to pmtk_search_pubmed().
}
