% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmed_get_affiliations.R
\name{pmed_get_affiliations}
\alias{pmed_get_affiliations}
\title{Download author/affiliation data from PubMed.}
\usage{
pmed_get_affiliations(pmids, cores = 3, ncbi_key = NULL, clean = T)
}
\arguments{
\item{pmids}{A vector of PMIDs}

\item{cores}{Numeric specifying number of cores to use}

\item{ncbi_key}{API key}

\item{clean}{Boolean}
}
\value{
A list of data frames
}
\description{
Download author/affiliation data from PubMed.
}
