% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmed_get_entities.R
\name{pmed_get_entities}
\alias{pmed_get_entities}
\title{Download PubTator Central NER annotations.}
\usage{
pmed_get_entities(pmids, cores = 3, verbose = T)
}
\arguments{
\item{pmids}{A vector of PMIDs}

\item{cores}{Numeric specifying number of cores to use}

\item{verbose}{Boolean}
}
\value{
A list of data frames
}
\description{
Download PubTator Central NER annotations.
}
