% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmed_get_icites.R
\name{pmed_get_icites}
\alias{pmed_get_icites}
\title{Extract citation info from NIH's iCite API.}
\usage{
pmed_get_icites(pmids, cores = 3, ncbi_key = NULL)
}
\arguments{
\item{pmids}{A character vector (of PMIDs)}

\item{cores}{An integer (specifying number of cores to use)}

\item{ncbi_key}{A character string (API key)}
}
\value{
A list of data frames
}
\description{
Extract citation info from NIH's iCite API.
}
