% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmed_get_neighbors.R
\name{pmed_get_neighbors}
\alias{pmed_get_neighbors}
\title{Measure cosine similarity.}
\usage{
pmed_get_neighbors(x, target, n = 10)
}
\arguments{
\item{x}{A matrix}

\item{target}{A character, or numeric vector.}

\item{n}{Integer}
}
\value{
A data frame.
}
\description{
Measure cosine similarity.
}
