% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmed_get_ortho.R
\name{pmed_get_ortho}
\alias{pmed_get_ortho}
\title{Measure JW spelling distance.}
\usage{
pmed_get_ortho(x, dict = mesh, term = "TermName", n = 5)
}
\arguments{
\item{x}{A character string}

\item{dict}{A data frame.}

\item{n}{Integer}

\item{TermName}{A character string}
}
\value{
A data frame.
}
\description{
Measure JW spelling distance.
}
