% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmed_get_data.R
\name{pmed_get_records}
\alias{pmed_get_records}
\title{Retrieve Data from PubMed Based on PMIDs}
\usage{
pmed_get_records(
  pmids,
  endpoint = c("pubtations", "icites", "affiliations", "pubmed", "pmc"),
  cores = 3,
  ncbi_key = NULL
)
}
\arguments{
\item{pmids}{A vector of PMIDs for which data is to be retrieved.}

\item{endpoint}{A character vector specifying the type of data to retrieve ('pubtations', 'icites', 'affiliations', 'pubmed', 'pmc').}

\item{cores}{Number of cores to use for parallel processing (default is 3).}

\item{ncbi_key}{(Optional) NCBI API key for authenticated access.}
}
\value{
A data.table containing combined results from the specified endpoint.
}
\description{
This function retrieves different types of data (like PubMed records, affiliations, iCites data, etc.) from PubMed based on provided PMIDs. It supports parallel processing for efficiency.
}
