% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmed_get_synonyms.R
\name{pmed_get_synonyms}
\alias{pmed_get_synonyms}
\title{Get synonyms from MeSH.}
\usage{
pmed_get_synonyms(x, mesh = pmed_tbl_mesh)
}
\arguments{
\item{x}{A character vector}

\item{mesh}{A data frame}
}
\value{
A character vector
}
\description{
Get synonyms from MeSH.
}
