% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmed_search_pubmed.R
\name{pmed_search_pubmed}
\alias{pmed_search_pubmed}
\title{Perform basic PubMed search.}
\usage{
pmed_search_pubmed(
  search_term,
  fields = c("TIAB", "MH"),
  verbose = TRUE,
  sleep = 1,
  retmax = 5e+06,
  is_pubmed_syntax = F
)
}
\arguments{
\item{search_term}{Query term as character string}

\item{fields}{PubMed fields to query}

\item{verbose}{Boolean}

\item{sleep}{In seconds.}

\item{retmax}{Max records}
}
\value{
A data frame of PMIDs
}
\description{
Perform basic PubMed search.
}
