% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmed_tbl_citations}
\alias{pmed_tbl_citations}
\title{Medline citation counts by year of publication.}
\format{
A data frame with 75 rows and 3 variables:
}
\source{
\url{https://www.nlm.nih.gov/bsd/medline_cit_counts_yr_pub.html}
}
\usage{
data(pmed_tbl_citations)
}
\description{
A dataset containing Medline citation counts by year from 1947-2020.
}
\keyword{datasets}
