% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmed_tbl_mesh}
\alias{pmed_tbl_mesh}
\title{MeSH hierarchical vocabulary & thesaurus.}
\format{
A data frame with 562300 rows and 10 variables:
}
\source{
\url{https://www.nlm.nih.gov/databases/download/mesh.html}
}
\usage{
data(pmed_tbl_mesh)
}
\description{
A dataset containing the MeSH hierarchical vocabulary & thesaurus.
}
\keyword{datasets}
