% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedr-package.R
\docType{package}
\name{pubmedr-package}
\alias{pubmedr}
\alias{pubmedr-package}
\title{pubmedr: PubMed Mining Toolkit}
\description{
An R package for querying the PubMed database & parsing retrieved records.
}
\keyword{internal}
