% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedr-package.R
\docType{package}
\name{pubmedtk-package}
\alias{pubmedtk}
\alias{pubmedtk-package}
\title{pubmedtk: PubMed Mining Toolkit}
\description{
An R package for querying the PubMed database & parsing retrieved records.
}
\keyword{internal}
