% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_one_pmid_metadata.R
\name{download_one_pmid_metadata}
\alias{download_one_pmid_metadata}
\title{Downloads metadata from Pubmed API for a single provided PMID and
exports}
\usage{
download_one_pmid_metadata(pmid, api_key)
}
\arguments{
\item{pmid}{A single PMID, e.g. "29559429"}

\item{api_key}{A valid Pubmed API key}
}
\description{
Downloads metadata from Pubmed API for a single provided PMID and
exports
}
