% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmed_search_pubmed.R
\name{pmed_search_pubmed}
\alias{pmed_search_pubmed}
\title{Perform basic PubMed search.}
\usage{
pmed_search_pubmed(query)
}
\arguments{
\item{query}{Query term as character string}
}
\value{
A vector of PMIDs
}
\description{
Perform basic PubMed search.
}
