% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedtk-package.R
\docType{package}
\name{pubmedtk-package}
\alias{pubmedtk}
\alias{pubmedtk-package}
\title{pubmedtk: 'Pubmed' Toolkit}
\description{
Provides various functions for retrieving and interpreting information from 'Pubmed' via the API, \url{https://www.ncbi.nlm.nih.gov/home/develop/api/}.
}
\details{
Pubmed toolkit
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bgcarlisle/pubmedtk}
  \item Report bugs at \url{https://github.com/bgcarlisle/pubmedtk/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Gregory Carlisle \email{murph@bgcarlisle.com} (\href{https://orcid.org/0000-0001-8975-0649}{ORCID})

}
\keyword{internal}
