% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_pubmed.R
\name{search_pubmed}
\alias{search_pubmed}
\title{Perform basic PubMed search.}
\usage{
search_pubmed(x)
}
\arguments{
\item{x}{Query term as character string}
}
\value{
A vector of PMIDs
}
\description{
Perform basic PubMed search.
}
